Python 3 improvement for the rdc.corr_plot and pcs.corr_plot user functions.
The world view is now set in floating point numbers.  In Python 2, the math.ceil() and math.floor()
functions return floats, whereas in Python 3 these functions return integers.  The behaviour is now
consistent in both Python versions, fixing a few system tests.

+++ pipe_control/pcs.py
@@ -500,8 +500,8 @@ def corr_plot(format=None, title=None, subtitle=None, file=None, dir=None, force
     size = len(data)
 
     # Round the data limits.
-    max_pcs = ceil(max_pcs)
-    min_pcs = floor(min_pcs)
+    max_pcs = float(ceil(max_pcs))
+    min_pcs = float(floor(min_pcs))
 
     # Only one data set.
     data = [data]
